+++ src/boids/gamestate.py
@@ -28,12 +28,14 @@ BINDINGS = [
 class GameState:
     """Interprets inputs and update internal game state to be read by simulation and renderer"""
 
-    def __init__(self):
-        print("Initialising Game State...")
-        self.key_bindings, self.state = self.build_key_binding(BINDINGS)
+    def __init__(self, quiet):
+        if not quiet:
+            print("Initialising Game State...")
+        self.key_bindings, self.state = self.build_key_binding(BINDINGS, quiet)
 
         # General
-        self.state["quit"]
+        self.state["quit"] = False
+        self.state["quiet"] = quiet
 
         # Simulation
         self.state["boids_count"] = None
@@ -43,12 +45,13 @@ class GameState:
         self.state["focus"] = None
         self.state["show_debug"] = "fps_cam_perf"
 
-    def build_key_binding(self, ctrl_list):
+    def build_key_binding(self, ctrl_list, quiet):
         """Build the key_bindings and state dictionaries from ctrl_list
         ctrl_list must be formatted as:
         [(control name, key, "pressed"|"held", default value), ...]
         """
-        print("Building Key Bindings...")
+        if not quiet:
+            print("> Building Key Bindings...")
         key_bindings = {"held": {}, "action": {}, "pressed": {}}
         state = {}